moved game carousel to new folder structure to better support fine-grained components
