+++ src/components/GameStage/GameCarousel/GameCarousel.css
