+++ src/components/GameStage/GameCarousel/GameCarousel.js
