+++ src/components/GameStage/GameDetail/GameDetails.js
